requirements.txt
requests
python-dotenv
supabase
web3
